+++ stm32/Src/ads.c
@@ -135,14 +135,14 @@ int ADC_filter(int readings[], int size){
 
 size_t ADC_measure(uint8_t *data) {
   // get timestamp
-  time_t ts = TimestampNow();
+  SysTime_t ts = SysTimeGet();
 
   // read voltage
   int adc_voltage = ADC_readVoltage();
   double adc_voltage_float = ((double) adc_voltage) / 1000.;
 
   // encode measurement
-  size_t data_len = EncodePowerMeasurement((uint32_t) ts, LOGGER_ID, CELL_ID,
+  size_t data_len = EncodePowerMeasurement(ts.Seconds, LOGGER_ID, CELL_ID,
                                            adc_voltage_float, 0.0, data);
 
   // return number of bytes in serialized measurement