+++ stm32/Src/fifo.c
@@ -24,10 +24,41 @@ static inline void update_addr(uint16_t *addr, const uint16_t num_bytes) {
   *addr = (*addr + num_bytes) % fram_buffer_size;
 }
 
+/**
+ * @brief Get the remaining space in the buffer
+ * 
+ * Calculates the difference between the write and read address in a single
+ * forward direction in the circular buffer. If they are equal then nothing has
+ * been written.
+ * 
+ * @return Remaining space in bytes
+ */
+static uint16_t get_remaining_space(void) {
+  uint16_t space_used = 0;
+  if (write_addr > read_addr) {
+    space_used = write_addr - read_addr;  
+  }
+  else if (write_addr < read_addr) {
+    space_used = fram_buffer_size - (read_addr - write_addr);
+  }
+  else {
+    // if anything is stored in buffer than entire capacity is used
+    // otherwise buffer is empty and all free space is available
+    if (buffer_len > 0) {
+      space_used = fram_buffer_size;
+    }
+    else {
+      space_used = 0;
+    }
+  }
+
+  uint16_t remaining_space = fram_buffer_size - space_used;
+  return remaining_space;
+}
+
 FramStatus FramPut(const uint8_t *data, const uint16_t num_bytes) {
   // check remaining space
-  uint16_t remaining_space = fram_buffer_size - (write_addr - read_addr);
-  if (num_bytes+1 > remaining_space) {
+  if (num_bytes+1 > get_remaining_space()) {
     return FRAM_BUFFER_FULL;
   }