+++ stm32/Src/main.c
@@ -123,8 +123,6 @@ int main(void)
   SensorsAdd(ADC_measure);
   //SensorsAdd(SDI12_Teros12Measure);
 
-  // start sensor measurements
-  SensorsStart();
 
   /* USER CODE END 2 */